Added ability to display all bounties on board index.
This is not the correct implementation. Currently @bounties has to be declared in the board controller to put the bounties in scope for the bounty template. This means the bounty controller, which is supposed to have all the restful actions of new, show, etc isn't being used at all. Gotta learn how to fix that.