Reintroduce a placeholder jQuery-UI theme (UI-Lightness) doing it right this time.
Note that in order to work with Rails' asset compilation, jQuery-UI
stylesheets needs to be modified to point to the right URLs for its
resources, mostly images. In script/ is a bash script called
prepare-jquery-ui-stylesheet.sh that modifies a jQuery-UI stylesheet to
be Rails-compatible. Run this script on any jQuery-UI stylesheet before
adding it to the project.