Build infrastructure for per-page JavaScript loading and use it for filters.
Check the "data-page" attribute of body with $("body").data("page") to
see what controller and action the page is being redered for. The data
will contain a string of the form "Controller#action", i.e.
"Bounties#index" for the main page.  Your JavaScript can check for the
controller or index it matches before executing any logic if it should
only be run on certain pages.